+++ app/src/main/AndroidManifest.xml
@@ -20,13 +20,13 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".SignUpStudentActivity"
+            android:name=".signup.SignUpStudentActivity"
             android:exported="false" />
         <activity
-            android:name=".LogInStudentActivity"
+            android:name=".login.LogInStudentActivity"
             android:exported="false" />
         <activity
-            android:name=".SignUpProfessorActivity"
+            android:name=".signup.SignUpProfessorActivity"
             android:exported="false">
         </activity>
         <activity
@@ -39,7 +39,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".LogInProfessorActivity"
+            android:name=".login.LogInProfessorActivity"
             android:exported="false" />
         <activity
             android:name=".nfc.reader.NFCReaderActivity"
