+++ app/src/main/java/com/example/urs_2024_25/login/LogInStudentActivity.java
@@ -1,4 +1,4 @@
-package com.example.urs_2024_25;
+package com.example.urs_2024_25.login;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,22 +10,24 @@ import android.widget.TextView;
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.urs_2024_25.nfc.reader.NFCReaderActivity;
+import com.example.urs_2024_25.R;
+import com.example.urs_2024_25.signup.SignUpStudentActivity;
 
-public class LogInProfessorActivity extends AppCompatActivity {
+public class LogInStudentActivity extends AppCompatActivity {
 
     private EditText loginEmail, loginPassword;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_log_in_professor);
+        setContentView(R.layout.activity_log_in_student);
 
-        loginEmail = findViewById(R.id.login_professor_email);
-        loginPassword = findViewById(R.id.login_professor_password);
-        TextView signupRedirectText = findViewById(R.id.signup_professor_redirect);
-        Button loginButton = findViewById(R.id.login_professor_button);
+        loginEmail = findViewById(R.id.login_student_email);
+        loginPassword = findViewById(R.id.login_student_password);
+        TextView signupRedirectText = findViewById(R.id.signup_student_redirect);
+        Button loginButton = findViewById(R.id.login_student_button);
 
         loginButton.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -39,15 +41,13 @@ public class LogInProfessorActivity extends AppCompatActivity {
                 if (pass.isEmpty()) {
                     loginPassword.setError("Password cannot be empty");
                 }
-
-                startActivity(new Intent(LogInProfessorActivity.this, NFCReaderActivity.class));
             }
         });
 
         signupRedirectText.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(LogInProfessorActivity.this, SignUpProfessorActivity.class));
+                startActivity(new Intent(LogInStudentActivity.this, SignUpStudentActivity.class));
             }
         });
     }
