+++ app/src/main/res/layout/activity_log_in_professor.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".LogInProfessorActivity">
+    tools:context=".login.LogInProfessorActivity">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
