+++ app/src/main/res/layout/activity_sign_up_student.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".SignUpStudentActivity">
+    tools:context=".signup.SignUpStudentActivity">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
